lb_submissions/SAI/RuBERT/requirements.txt
torch==1.9.0
torchtext==0.6.0
tensorflow==2.6.0
keras==2.6.0
pandas==1.3.5
transformers==4.12.5
click==7.1.2
nltk==3.4.5
sklearn-crfsuite==0.3.6
